add external library: opus speex update external library: zlib libpng lcore: modify hashmap string lmath: add AABB Coordinate lgraphics: linear color space support, add StructuredBuffer lframework: modify SceneManager lrenderlib: add Dipole Model SSS Integrator add sound library lsound

+++ lframework/scene/ConstantBufferTable.cpp
@@ -71,7 +71,7 @@ namespace
             return allocateLarge(size);
         }
 
-        u16 index = calcIndex(size);
+        u16 index = static_cast<u16>(calcIndex(size));
         LASSERT(0<=index && index<TableSize);
 
         TableEntry& tableEntry = entries_[index];
